# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BanglaConnect API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /blogs:
    get:
      operationId: listBlogs
      tags: [blogs]
      summary: List blog posts
      parameters:
        - name: tag
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of blog posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Blog"
    post:
      operationId: createBlog
      tags: [blogs]
      summary: Create a blog post
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogInput"
      responses:
        "201":
          description: Created blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Blog"
  /blogs/{id}:
    get:
      operationId: getBlog
      tags: [blogs]
      summary: Get a blog post by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Blog"
        "404":
          description: Not found
    put:
      operationId: updateBlog
      tags: [blogs]
      summary: Update a blog post
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogInput"
      responses:
        "200":
          description: Updated blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Blog"
    delete:
      operationId: deleteBlog
      tags: [blogs]
      summary: Delete a blog post
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── News ───────────────────────────────────────────────────────────────────
...
  /admin/mentors:
    get:
      operationId: listPendingMentors
      tags: [admin]
      summary: List mentors pending approval
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of mentor profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Profile"
  /admin/mentors/{id}/status:
    patch:
      operationId: updateMentorStatus
      tags: [admin]
      summary: Approve or reject a mentor
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MentorStatusUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /admin/tickets:
    get:
      operationId: listAllSupportTickets
      tags: [admin]
      summary: List all support tickets (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of support tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupportTicket"
  /admin/tickets/{id}:
    patch:
      operationId: updateTicketStatus
      tags: [admin]
      summary: Update support ticket status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketStatusUpdate"
      responses:
        "200":
          description: Updated ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicket"
  /admin/users:
    get:
      operationId: listAllUsers
      tags: [admin]
      summary: List all users (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of all user profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Profile"

  # ── Chat ───────────────────────────────────────────────────────────────────
  /chat/conversations:
    get:
      operationId: listConversations
      tags: [chat]
      summary: List current user's conversations
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: startConversation
      tags: [chat]
      summary: Start or retrieve a conversation with a user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartConversationInput"
      responses:
        "200":
          description: Conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /chat/conversations/{id}/messages:
    get:
      operationId: listMessages
      tags: [chat]
      summary: List messages in a conversation
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send a message in a conversation
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageInput"
      responses:
        "201":
          description: Sent message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Profile:
      type: object
      required: [id, clerkId, role, fullName, country, mentorStatus, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [student, mentor, professional, admin]
        bio:
          type: ["string", "null"]
        country:
          type: string
        city:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        interests:
          type: array
          items:
            type: string
        professionalField:
          type: ["string", "null"]
        profession:
          type: ["string", "null"]
        specialties:
          type: array
          items:
            type: string
        mentorAvailable:
          type: boolean
        mentorStatus:
          type: string
          enum: [pending, approved, rejected]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Blog:
      type: object
      required: [id, title, content, authorId, tags, published, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        excerpt:
          type: ["string", "null"]
        coverImageUrl:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        authorId:
          type: integer
        author:
          $ref: "#/components/schemas/Profile"
        published:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
    BlogInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        content:
          type: string
        excerpt:
          type: string
        coverImageUrl:
          type: string
        tags:
          type: array
          items:
            type: string
        published:
          type: boolean

    # ── News ─────────────────────────────────────────────────────────────────
...
    SupportTicket:
      type: object
      required: [id, userId, subject, description, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        subject:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, in_progress, resolved]
        adminNotes:
          type: ["string", "null"]
        user:
          $ref: "#/components/schemas/Profile"
        createdAt:
          type: string
        updatedAt:
          type: string
...
    TicketStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [open, in_progress, resolved]
        adminNotes:
          type: string
    MentorStatusUpdate:
      type: object
      required: [mentorStatus]
      properties:
        mentorStatus:
          type: string
          enum: [pending, approved, rejected]

    # ── Chat ──────────────────────────────────────────────────────────────────
    Conversation:
      type: object
      required: [id, participant1Id, participant2Id, lastMessageAt, createdAt]
      properties:
        id:
          type: integer
        participant1Id:
          type: integer
        participant2Id:
          type: integer
        participant1:
          $ref: "#/components/schemas/Profile"
        participant2:
          $ref: "#/components/schemas/Profile"
        lastMessage:
          $ref: "#/components/schemas/ChatMessage"
        lastMessageAt:
          type: string
        createdAt:
          type: string
    StartConversationInput:
      type: object
      required: [otherProfileId]
      properties:
        otherProfileId:
          type: integer
    ChatMessage:
      type: object
      required: [id, conversationId, senderId, content, createdAt]
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        senderId:
          type: integer
        sender:
          $ref: "#/components/schemas/Profile"
        content:
          type: string
        readAt:
          type: ["string", "null"]
        createdAt:
          type: string
    SendMessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string