# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BanglaConnect API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProfileMentorStatus } from './profileMentorStatus';
import type { ProfileRole } from './profileRole';

export interface Profile {
  id: number;
  clerkId: string;
  fullName: string;
  role: ProfileRole;
  /** @nullable */
  bio?: string | null;
  country: string;
  /** @nullable */
  city?: string | null;
  skills?: string[];
  interests?: string[];
  /** @nullable */
  professionalField?: string | null;
  /** @nullable */
  profession?: string | null;
  specialties?: string[];
  mentorAvailable?: boolean;
  mentorStatus: ProfileMentorStatus;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}