# lib/api-zod/src/generated/types/supportTicket.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BanglaConnect API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Profile } from './profile';
import type { SupportTicketStatus } from './supportTicketStatus';

export interface SupportTicket {
  id: number;
  userId: number;
  subject: string;
  description: string;
  status: SupportTicketStatus;
  /** @nullable */
  adminNotes?: string | null;
  user?: Profile;
  createdAt: string;
  updatedAt: string;
}